optimization

+++ src/main/java/com/wjw/blog/controller/admin/TypeController.java
@@ -2,7 +2,9 @@ package com.wjw.blog.controller.admin;
 
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
+import com.wjw.blog.dto.BlogShow;
 import com.wjw.blog.entity.Type;
+import com.wjw.blog.service.BlogService;
 import com.wjw.blog.service.TypeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -19,12 +21,16 @@ public class TypeController {
     @Autowired
     private TypeService typeService;
 
+    @Autowired
+    private BlogService blogService;
+
     @GetMapping("/types")
     public String types(Model model, @RequestParam(value = "p", defaultValue = "1") Integer pageNum) {
         PageHelper.startPage(pageNum, 5);
         List<Type> allTypes = typeService.getAdminType();
         PageInfo<Type> pageInfo = new PageInfo<>(allTypes);
         model.addAttribute("pageInfo", pageInfo);
+        model.addAttribute("types", typeService.getAllType());
         return "admin/types";
     }
 
@@ -67,6 +73,11 @@ public class TypeController {
 
     @GetMapping("/types/{id}/delete")
     public String delete(@PathVariable("id") Long id, RedirectAttributes attributes) {
+        List<BlogShow> blogs = blogService.getAllByTypeId(id);
+        if(blogs.size() != 0) {
+            attributes.addFlashAttribute("message", "该分类下仍存在博客， 请修改博客分类后进行删除");
+            return "redirect:/admin/types";
+        }
         typeService.deleteType(id);
         attributes.addFlashAttribute("message", "删除成功");
         return "redirect:/admin/types";
